+++ test.py
@@ -1,0 +1,6 @@
+from sqlalchemy.sql import select
+
+
+cte = select([]).cte()
+
+reveal_type(cte)