+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Changes to be tagged
 
+## 5.37.4 - 2020-01-30
+### Added
+- `battleCharaOffsetZ` to raw servant's `mstSvtLimitAdd`.
+
+### Changed
+- `scale` in raw servant's `mstSvtScript` to have decimal type instead of integer.
+
 ## 5.37.3 - 2020-01-28
 ### Fixed
 - Mash should return the same NP but with different priorities now. Previously, nice Mash returned the same NP 800105 with the same priority 105. Now there should be two 800105 NPs with priorities 105 and 106.